Various fixes for buildrelease script contributed by a.c.li
(1) Guess the location of koha-html in a hopefully more sensible way

(2) Use getpwuid() to figure where root's home directory is; traditionally /root
is usually not right

(3) Check for the return code of chdir calls

(4) Removed compilation warning messages (replaced \RC with explicit use of
concatenation operator)

(5) The questions about tagging and updates could be quite confusing to the new
user (of buildrelease); I tried to reword a couple of sentences a bit, and
rearranged their orders a bit, to reduce the likelihood of getting confused.
Needless questions need not be asked (confirming whether to tag when the user
says he/she doesn't want to tag).

(6) Don't say it is tagging the CVS files when it is not.