Road to 1.3.2 * bugfixes and improvements * manage mandatory MARC subfields * new table : authorised_values. this table contains categories and authorised values for the category. On MARC management, you can map a subfield to a authorised_values category. If you do this, the subfield can only be filled with a authorised_value of the selected category. this submit contains everything needed : * updatedatabase * admin screens * "links" management * creation of a html-list if a subfield is mapped to an authorised value.
Note this is different from authorities support, which will come soon.
The authorised_values is supposed to contains a "small" number of authorised values for a category (less than 50-100). If you enter more authorised values than this, it should be hard to find what you want in a BIG list...