This should fix bug 734.
Added a test for the "zero rows affected" condition when trying to update
aqorderbreakdown. According to DBI(3), execute() returns the "number of
rows affected", so we know whether the update has succeeded, or whether we
need to redo the update as an insert.

The error condition (return value -1) is still not being handled. But the
old code did not check for errors either, so this would be reasonable for
the time being.