new feature : online HELP On each template, a "HELP" link has been added. it opens a popup, filled with an help template. the help template is in the help/ directory. The help.pl files builds the online help using the name of the caller page (referer).
If the user clic on "help" when on page "admin/itemtypes.pl", the "help/admin/itemtypes.tmpl" is opened.
If the "help/admin/itemtypes.tmpl" does not exists, a default "sorry, no help on this topic" page is shown.

Works fine, but now, the big work has to be done : create the online help pages...