Taught buildrelease what the misc directory and release 2.2 is, and also that version numbers might only have 1 dot. This should fix bugs 889, 890.
A --verbose or -v option has been added in the process of doing the above.

It now chdir to .. if it discovers that it is being run inside misc, since
running it inside misc screws up various things, and I never remember
whether it should be run inside misc or not.

A confirmation message about whether to tag or not is now displayed after
the user answers the question about tagging. I have never felt safe not
having the release builder confirm with me that I really am not tagging
anything.