This is currently a copy of C4::Date
But C4::Date uses Date::Manip, which in the authors own words

"If you look in CPAN, you'll find that there are a number of Date and Time packages.  Is Date::Manip the one you should be using?  In my
opinion, the answer is no most of the time."

He goes on to say, that because Date::Manip is powerful and is written fully in perl its also slow.
Now Circulation needs to be as fast as possible. And C4::Date isnt actually doing anything particularly tricky,
So im working on C4::Circulation::Date to be a replacement, in an attempt to win some speed