VERY IMPORTANT NOTES!!! READ OR YOU MAY RUIN YOUR ENCODING
Previous commits replaced the use of the MARChtml2marc
routine with a MARChtml2xml that reads in the form data and
converts it to a MARC::File::XML object. It then passes that
to MARC::Record via the new_from_xml() method.

This works perfectly and fixes all of the bugs with the old
MARChtml2marc routine (see mail to koha-devel). However,
there is one gotcha. The MARC::File::XML new_from_xml() method
converts UTF-8 data into MARC-8 by default for historical reasons.
This means that entry of characters outside the normal ascii range
in UTF-8 will come out as gibberish in MARC-8. Never fear, there
is a solution.

This commit uses code written for MARC::File::XML version 0.82 to
add the ability to specify the encoding desired (I chose UTF-8
because I can't think of a reason to ever need MARC-8)

The bottom line is that to properly encode records YOU WILL HAVE
TO UPDATE MARC::File::XML to at least CPAN verion 0.82 !!!!!