IMPORTANT:
MARC::Record objects coming in from the reservoir are probably going to
be encoded as MARC-8. This enures that they will be converted into UTF-8
before being displayed to the user (if they are in fact MARC-8). It's most
easily done by turning the record into a MARC::File::XML object and then
going back to MARC::Record -- easiest to do for me that is. Of course, a
better solution is to create a routine that changes the encoding of a
MARC::Record object (I was surprised to find such did not exist). I have
tested this method for MARC21, but I'm not familiar with the issues with
encoding in UNIMARC: are UNIMARC records ever encoded in charater sets
other than MARC-8 and UTF-8? (if so, I will need to adjust this code).