slight change in receiveorder() call, added bookfund arg. bookfund now updated during receive for HLT.
Mason's ACQ fixes from HLT-DEV , work in progress.

Adding code to check existing orders in basket to avoid duplication of rows.

Addorder() now allows the user to select whether: the bibitem for the order will be an existing bibitem, or a newly created  bibitem.

adding back in the part the saves the edit page's changes.

Fix for bug 3589, can now order and receive donations in one go.

Fixing a bug, acquisitions wasnt making a new biblioitem when needed.