[IMPORTANT]
when cloning a tag, there was a problem to detect the tag change when rebuilding the XML/MARC record :
the test was done on an empty field.
with the new CloneTag sub, this empty field was quite tricky to create on the fly.
so i've added on each subfield a <input name="firstsubfield"> saying wether the subfield is the 1st or not.
Thus, when a tag is cloned, it is cloned properly.

HTML2marc and well as HTML2xml have been modified, but PLEASE DOUBLE CHECK my work