BugFixing : MARCdetail : displayin field values with ESCAPE=HTML  (in order to manage  '<''>' characters)
Adding  Hierarchy display for authorities.
Please Note That it relies on the fact that authorities id are stored in $3 of authorities notice.
And Broader terms is supposed to be indicated by a g for 550$5 subfield, narrower term : an h for the same subfield.

It CAN SURELY be generalised but only with a bunch of sytem preferences.

I added the ability to do a search on ANY authtypecode.