BugFixing 1124 1125. 1125 is a bug that seems to me a design error.
To manage all the serials, it may be better to display the list as in full-serial-issue.tmpl and allow to change through a click on the issue. Receiving the next one or a manual one would be in a different part of the screen, short history (missing list and receivedlist) too.