Bug Fixing : Adding item had a javascript Error.
So.
There are no more errors on saving an item.
There are still little problems when you duplicate a subfield :
Display table is not adapted
When you reedit, all the duplicate field values are concatenated.
But apart from that, it works.
Donot have time enough to work on it these days.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>