bump to 3.00.00.017 : adding columns to zebraqueue
The zebraque_start scripts works, but it not friendly when you try to track what he did
as records are deleted after done.
This commit changes the behaviour : 2 columns are added done & timestamp.
done is set to 1 when a line have been done. And the timestamp contains the timestamp of the last action (either line creation if done=0 or operation if done=1)

should be helpfull to track problem.

the table will grow, but i'll add soon a DELETE FROM zebraqueue WHERE timestamp > 30 days or something like that

Signed-off-by: Chris Cormack <crc@liblime.com>
Signed-off-by: Joshua Ferraro <jmf@liblime.com>