viewlog.tmpl + calendar.inc: builds Date_from_syspref javascript function.
This is likely the durable fix to bridge the gap between syspref on the server
and client side validation scripts.  It may be useful to separate Date_from_syspref
from the calendar js, but at this time their applicability is 1-to-1.
Note also js debug variable and CGI param('debug') responsiveness.

Signed-off-by: Chris Cormack <crc@liblime.com>
Signed-off-by: Joshua Ferraro <jmf@liblime.com>