Important Commit For UNIMARC Items.
Nonetheless, you are right to consider items table as the reference.
We need one, and we agree on this principle.

What is done with this commit is :
- adding support for marc subfields not linked to items table.

+ removing some $dbh parameters in some function signature.
I think this is rel2_2 style programming and may be avoided using C4::Context->dbh.
Or there is a reason using dbh parameter that you know and not me.

Signed-off-by: Chris Cormack <crc@liblime.com>
Signed-off-by: Galen Charlton <galen.charlton@liblime.com>
Signed-off-by: Joshua Ferraro <jmf@liblime.com>