partial fix for bug 1417
The Notes field should now be displayed correctly
when editing an order - the query in GetOrder
was modified to have aqorder's notes column
appear last, which means that it is the one
selected for the data hash (instead of biblio.notes
or biblioitems.notes).

This is an ugly fix - GetOrder needs to be
refactored so that it doesn't depend on
selecting all columns from several different
tables.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>