patron attributes - OPAC display
If ExtendedPatronAttributes is ON, any patron attributes
whose type is marked as displayble in the OPAC are now
displayed in the 'my details' page in the OPAC (assuming
that the patron is logged in).  These appear in the
'Library Use' section.

Currently, it is not possible for a patron to suggest
changes to any of their extended attributes using
opac-userupdate.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>