patron attributes - staff search and display
If ExtendedPatronAttributes is ON, patron attributes
can be displayed and edited by staff members.

The patron attributes are displayed and edited
in a new section (step 4) of the patron details page.  Each
attribute is identified by its type (including type code
and type description) and value (included value description
if the attribute is controlled by an authorised value category).

Note: if, for a repeatable attribute type, the staff member
enters multiple copies of that type with the same value, duplicates
are removed when the patron record is saved.  Also, if the type
is repeatable, a JavaScript link allows the staff member to
create new attributes of that type.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>