MARC21 frameworks: keep a tag's subfields in the same tab
Changed default mandatory and optional MARC21 frameworks
to ensure that all the subfields of 037, 100, 110, 111, 130,
240, 243, 400, 410, 411, 440, 584, and 760 are all either
ignored (tab = -1) or assigned to the same editing tab.

Most of these changes involve the subfield $9 in authority-controlled
fields, which for the affected 1xx, 2xx, and 4xx fields was assigned
to tab 6, not 1, 2, or 4.  Because of this, any such field with a
$9 was effectively split, leading to record corruption.

This is part of the fix for bug 2017.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>