kohabug 2205 - fix 404 when traversing pagination link in OPAC authorities search
Currently, the heading sort order is a required parameter, but was
not supplied to the template or used in the search results.  As
a result of this patch, clicking a link to view another page
of the authority search results will now work.

Outstanding issue not addressed by this patch: when searching
authorities in both the staff interface and the OPAC, choosing
sort order "none" results in a crash.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>